fix: remove geth logger to utils
